SameDiff reorganization - round 2
